backend/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.2
boto3==1.33.10
botocore==1.33.10
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.0.1
click==8.1.7
constantly==15.1.0
crontab==1.0.1
cryptography==44.0.1
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==4.2.25
django-allauth==65.4.1
django-cors-headers==3.14.0
django-filter==23.1
django-rq==3.0.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
freezegun==1.5.1
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==23.0.0
hvac==2.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
promise==2.3
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pynamodb==5.5.1
pyOpenSSL==24.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rq==2.1.0
rq-scheduler==0.14.0
s3transfer==0.8.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.5.0
stripe==10.5.0
text-unidecode==1.3
tomli==2.0.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.6.0
urllib3==1.26.18
zope.interface==5.5.2
